third_party/blink/web_tests/fast/css/readonly-pseudoclass-opera-003-expected.txt

:read-only doesn't apply to following controls. Fails if FAILURE is shown somewhere.

0: SUCCESS
1: SUCCESS
2: SUCCESS
3: SUCCESS
4: SUCCESS
5: SUCCESS
6: SUCCESS
7: SUCCESS
8: SUCCESS
9: SUCCESS
